+++ src/renderer/store/actions/PlaylistsActions.ts
@@ -3,8 +3,8 @@ import electron from 'electron';
 import * as m3u from 'm3ujs';
 
 import { Playlist, TrackModel, PlaylistModel } from '../../../shared/types/museeks';
+import history from '../../lib/history';
 import store from '../store';
-import history from '../../router/history';
 import types from '../action-types';
 import * as app from '../../lib/app';
 import * as ToastsActions from './ToastsActions';
